fix possible implicit conversion issues
count_quote_chars was returning a size_t,
but  its return value was often stored in an int.
since an int is large enough here, I converted the return value to int.

+++ inc/utils.h
@@ -90,7 +90,7 @@ int			complete_quote(t_shell *shell, char **word);
 **	returns:
 **	number of quote characters in the string passed as argument.
 */
-size_t		count_quote_chars(char *str);
+int			count_quote_chars(char *str);
 
 /*
 **	makes a copy of str with a backslash ([\]) in front of every
